pyproject.toml
[project]
name = "swingtraderai"
version = "0.1.0"
description = ""
authors = [
	{name = "Bulat Nasyrov",email = "bulatnasirov2003@gmail.com"}
]
license = {text = "Apache2.0"}
readme = "README.md"
requires-python = ">=3.12,<=3.14"
dependencies = [
	"fastapi>=0.115.0",
	"uvicorn[standard]>=0.30.0",

	"sqlalchemy[asyncio]>=2.0.0",
	"asyncpg>=0.29.0",
	"alembic>=1.13.0",

	"pydantic[email] (>=2.12.5,<3.0.0)",
	"pydantic-settings>=2.5.0",

	"aiogram>=3.13.0",

	"yfinance>=0.2.40",
	"ccxt>=4.3.0",
	"pandas>=2.2.0",
	"pandas-ta>=0.3.14b0",
	"numpy>=2.0.0",

	"apscheduler>=3.10.0",

	"langchain>=0.3.0",
	"langgraph>=0.2.0",

	"python-dotenv>=1.0.0",
	"httpx>=0.27.0",
	"tenacity>=9.0.0",
	"celery (>=5.6.2,<6.0.0)",
	"redis (>=7.2.1,<8.0.0)",
	"passlib (>=1.7.4,<2.0.0)",
	"python-jose (>=3.5.0,<4.0.0)",
	"cryptography (>=46.0.5,<47.0.0)",
	"python-multipart (>=0.0.22,<0.0.23)",
	"bcrypt (==4.2.1)",
]

packages = [
	{ include = "swingtraderai" }
]

[tool.poetry.group.test.dependencies]
black = "^24.10.0"
ruff = "^0.9.0"
mypy = "^1.14.0"
pytest = "^8.0.0"
pre-commit = "^4.5.1"
pytest-asyncio = "^1.3.0"
pytest-cov = "^7.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.format]
indent-style = "tab"
quote-style = "double"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "C4"]
ignore = [
	"W191",
	"B008",
]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]
exclude = ["tests/.*", "alembic/.*"]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true

[[tool.mypy.overrides]]
module = ["tests.*", "alembic.*"]
ignore_errors = true
disallow_untyped_defs = false
check_untyped_defs = false

[tool.pytest.ini_options]
minversion = "8.0"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

testpaths = [
	"tests",
]

python_files = "test_*.py *_test.py"
python_functions = "test_*"

addopts = [
	"--strict-markers",
	"--tb=short",
	"-ra",
	"--showlocals",
	"--asyncio-mode=auto",
]

markers = [
	"asyncio: mark test as async (autouse with pytestmark)",
	"slow: marks tests as slow (deselect with --no-slow)",
	"db: tests that need real database",
]

filterwarnings = [
	"ignore::DeprecationWarning:sqlalchemy",
	"ignore:'crypt' is deprecated:DeprecationWarning",
]

norecursedirs = [
	".git",
	".venv",
	"migrations",
	"__pycache__",
]
